feat(onboarding): Implement CAS Parser and PDF Upload
This commit introduces the ability for users to upload Consolidated Account Statement (CAS) PDF files during the onboarding flow to automatically populate their investment data. It also includes several refinements to the login process, navigation, and data models across the application.

**Key Changes:**

*   **CAS Parser Integration:**
    *   **`CASParserViewModel.kt`:** A new ViewModel to handle the selection, password management, and uploading of CAS PDF files.
    *   **`CASRepo.kt` & `CASReportRepo.kt`:** Added a new repository and interface to interface with the CAS parser API (`api.casparser.in`).
    *   **`CASUploadScreenDialog.kt`:** A new dialog UI that allows users to pick a PDF, enter its password, and view the upload progress.
    *   **`CASResponseDto.kt` & `Summary.kt`:** New data transfer objects to handle the structured response from the CAS parser, including demat, mutual funds, insurance, and NPS values.

*   **Platform-Specific Document Management:**
    *   **`PdfPicker.kt` (Common):** Defined expected/actual declarations for `DocumentManager` and `SharedDocument`.
    *   **`PdfManagerAndroid.kt` & `PdfManagerIOS.kt`:** Implemented native document picking logic using `ActivityResultContracts.OpenDocument` on Android and `UIDocumentPickerViewController` on iOS.

*   **Onboarding Enhancements:**
    *   **`OnBoardingRootScreen.kt`:** Integrated the `CASUploadScreenDialog`. Successfully parsed data now automatically updates Stock, Insurance, and Mutual Fund asset fields.
    *   **`GoalScreenViewModel.kt`:** Added validation logic for goal ages and years (e.g., ensuring retirement age is greater than current age and less than life expectancy).
    *   **`GoalMapper.kt`:** Added mappers to transform domain goal requests into DTOs and JSON elements for API submission.
    *   **`ConfirmationScreen.kt`:** Now displays a summary of added goals and includes a loading state for the "Confirm" button.

*   **Authentication & Profile:**
    *   **Sign-Out Implementation:** Added `signOut()` to `UserAuthenticationRepo` and integrated it into the `ProfileScreen.kt` UI.
    *   **Login Flow:** Fixed OTP timer issues in `LoginScreenViewModel.kt` and updated `OtpScreen.kt` to allow manual OTP resending.
    *   **`AuthMappers.kt` & `Metadata.kt`:** Renamed `is_onboarding_complete` to `is_onboarding_completed` to match the backend schema.

*   **Networking & Utilities:**
    *   **`HttpClient.kt`:** Improved bearer token loading logic to handle null/empty tokens gracefully.
    *   **`PermissionManager.kt`:** Introduced a cross-platform permission management system for Gallery and Document access.
    *   **`PdfDownloader.kt` (Android):** Enhanced with `NotificationManager` support to show a system notification when a PDF download completes, allowing users to open the file directly.

*   **Data Model Refinements:**
    *   Updated `ComputedMetrics.kt`, `Projection.kt`, and `GoalHit.kt` to use `Long` or `Double` for financial values instead of `Int` to prevent overflow and support decimal precision.

*   **Navigation:**
    *   Updated `AppNavigation.kt` and `BaseNavigation.kt` to support the sign-out flow, ensuring the stack is cleared when returning to the login screen.

*   **Android Manifest & Resources:**
    *   Added `POST_NOTIFICATIONS` permission and a `FileProvider` to support PDF downloading and viewing.
    *   Added `file_paths.xml` for secure file sharing.

+++ composeApp/src/commonMain/kotlin/org/sharad/velvetinvestment/data/remote/model/firereport/PortfolioValue.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class PortfolioValue(
-    val emi_exclude: Int,
-    val emi_include: Int
+    val emi_exclude: Long,
+    val emi_include: Long
 )
